File rename: "ExpoentialSpringForce" to "ExponentialContact"
This name change made the name shorter and will help distinguish the OpenSim classess (ExponentialContact) from the Simbody classes (ExponentialSpringForce).
