+++ OpenSim/Simulation/Model/ExponentialContact.h
